requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu117
albumentations==1.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
filelock==3.12.2
idna==3.4
imageio==2.31.1
Jinja2==3.1.2
joblib==1.3.1
lazy_loader==0.3
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
opencv-python==4.8.0.74
opencv-python-headless==4.8.0.74
packaging==23.1
Pillow==10.0.0
PyWavelets==1.4.1
PyYAML==6.0.1
qudida==0.0.4
requests==2.31.0
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.1
sympy==1.12
threadpoolctl==3.2.0
tifffile==2023.7.18
torch==2.0.1
torchaudio==2.0.2
torchsummary==1.5.1
torchvision==0.15.2
tqdm==4.65.0
typing_extensions==4.7.1
urllib3==2.0.4
omegaconf
fvcore
iopath
submitit
torchmetrics==0.10.3
--extra-index-url https://pypi.nvidia.com
cuml-cu11
